java/src/test/java/com/yourtion/leetcode/easy/c9025/SolutionTest.java
package com.yourtion.leetcode.easy.c9025;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("面试题 08.01. 三步问题")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments(3, 4),
                arguments(5, 13)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void waysToStep(int source, int res) {
        System.out.printf("runTest: %d , res: %d", source, res);
        Assertions.assertEquals(res, new Solution().waysToStep(source));
    }
}